move EditorialSupportTeams model into models
