+++ app/views/editorialSupportStatus.scala.html
@@ -1,4 +1,4 @@
-@(teams: List[model.EditorialSupportTeam])
+@(teams: List[EditorialSupportTeam])
 
 @layout("Support") {
     <div class="admin">
